Moving event tests to a common file
